homeFeed complete

+++ functions/src/controllers/posts.controller.ts
@@ -1,7 +1,6 @@
 import * as admin from "firebase-admin";
 import * as config from "../config/config.json";
 import { Storage } from "@google-cloud/storage";
-import { Request, Response, NextFunction } from "express";
 const formParser = require("../utils/formParser");
 import { Post } from "../models/Post";
 const { v4: uuidv4 } = require("uuid");
@@ -12,6 +11,7 @@ const suid = new ShortUniqueId();
 
 export let db = admin.firestore();
 const postCollection = "posts";
+const homeFeedCollection = "homeFeed";
 
 const storage = new Storage({
   projectId: config.project_id
@@ -20,11 +20,7 @@ const storage = new Storage({
 
 const bucket = storage.bucket(`${config.project_id}.appspot.com`);
 
-export let createPost = async (
-  req: Request,
-  res: Response,
-  next: NextFunction
-) => {
+export let createPost = async (req,res,next) => {
   try {
     const formData = await formParser.parser(req, MAX_SIZE);
     const files = formData.files;
@@ -54,7 +50,7 @@ export let createPost = async (
         imageUrls.push(publicUrl);
         if (imageUrls.length === noOfImages) {
           let post: Post = {
-            user_id: formData["user_id"],
+            user_id: req.user.id,
             imageUrl: imageUrls,
             comment: [],
             likers: [],
@@ -63,9 +59,22 @@ export let createPost = async (
             Timestamp: new Date()
           };
           const newDoc = await db.collection(postCollection).add(post);
-          return res
-            .status(200)
+          const newId = newDoc.id;
+          res.status(200)
             .json({ message: "Post created successfully!", id: newDoc.id });
+
+          //For homeFeed  
+          const userDoc = await db.collection("users").doc(req.user.id).get();
+          const userConnectionsId = userDoc.data().connections;
+          await db.collection(homeFeedCollection).doc(req.user.id).collection("feed").doc(newId).set({
+            ...post
+          })
+          userConnectionsId.forEach(id=>{
+            db.collection(homeFeedCollection).doc(id).collection("feed").doc(newId).set({
+              ...post
+            })
+          })
+          return;
         }
       });
       blobWriter.end(file.content);
@@ -76,15 +85,11 @@ export let createPost = async (
   }
 };
 
-export let getPosts = async (
-  req: Request,
-  res: Response,
-  next: NextFunction
-) => {
+export let getPosts = async (req,res,next) => {
   //user_id==id of author
   try {
-    const formData = await formParser.parser(req, MAX_SIZE);
-    const user_id = formData["user_id"];
+    //const formData = await formParser.parser(req, MAX_SIZE);
+    const user_id = req.user.id;
     console.log(user_id);
     await db
       .collection(postCollection)
@@ -110,11 +115,7 @@ export let getPosts = async (
   }
 };
 
-export let updatePost = async (
-  req: Request,
-  res: Response,
-  next: NextFunction
-) => {
+export let updatePost = async (req,res,next) => {
   try {
     const formData = await formParser.parser(req, MAX_SIZE);
     const files = formData.files;
@@ -152,7 +153,7 @@ export let updatePost = async (
             imageUrls.push(publicUrl);
             if (imageUrls.length === files.length) {
               let post: Post = {
-                user_id: formData["user_id"],
+                user_id: req.user.id,
                 imageUrl: imageUrls,
                 comment: [],
                 likers: [],
@@ -163,7 +164,7 @@ export let updatePost = async (
               db.collection(postCollection)
                 .doc(formData["post_id"])
                 .update(post)
-                .then(() => {
+                .then( async () => {
                   res
                     .status(200)
                     .json({ message: "Post Update successfully!" });
@@ -173,6 +174,19 @@ export let updatePost = async (
                     const img = bucket.file(imageName);
                     img.delete();
                   });
+
+                  //feed update
+
+                  const userDoc = await db.collection("users").doc(req.user.id).get();
+                  const userConnectionsId = userDoc.data().connections;
+                  await db.collection(homeFeedCollection).doc(req.user.id).collection("feed").doc(formData["post_id"]).update({
+                    post
+                  });
+                  userConnectionsId.forEach(id=>{
+                  db.collection(homeFeedCollection).doc(id).collection("feed").doc(formData["post_id"]).update({
+                      post
+                    })
+                  })
                   return;
                 });
             }
@@ -186,37 +200,41 @@ export let updatePost = async (
   }
 };
 
-export let deletePost = async (
-  req: Request,
-  res: Response,
-  next: NextFunction
-) => {
+export let deletePost = async (req,res,next) => {
   const formData = await formParser.parser(req, MAX_SIZE);
   let postId = formData["postId"];
-  await db
-    .collection(postCollection)
-    .doc(postId)
-    .delete()
-    .then(() => {
-      res.status(200).json({ message: "Post deleted successfully" });
-      return;
-    })
-    .catch(error => {
-      console.log(error);
-      return res.status(400).json({ message: "Something went Wrong!" });
-    });
+  const authorId = (await db.collection(postCollection).doc(postId).get()).data().user_id;
+  if(authorId === req.user.id){
+    await db
+      .collection(postCollection)
+      .doc(postId)
+      .delete()
+      .then(async () => {
+        res.status(200).json({ message: "Post deleted successfully" });
+        
+        const userDoc = await db.collection("users").doc(req.user.id).get();
+          const userConnectionsId = userDoc.data().connections;
+          await db.collection(homeFeedCollection).doc(req.user.id).collection("feed").doc(postId).delete();
+          userConnectionsId.forEach(async id=>{
+              await db.collection(homeFeedCollection).doc(id).collection("feed").doc(postId).delete();
+        })
+        return;
+      })
+      .catch(error => {
+        console.log(error);
+        return res.status(400).json({ message: "Something went Wrong!" });
+      });
+  }else{
+    return res.status(400).json({ message: "You are not authorized to Delete this Post!!" });
+  }
 };
 
-export let likePost = async (
-  req: Request,
-  res: Response,
-  next: NextFunction
-) => {
+export let likePost = async (req,res,next) => {
   try {
     const formData = await formParser.parser(req, MAX_SIZE);
     const postId = formData["postId"];
     //here userId is the id of that user who is liking this post
-    const userId = formData["userId"];
+    const userId = req.user.id;
     db.collection(postCollection)
       .doc(postId)
       .get()
@@ -248,16 +266,12 @@ export let likePost = async (
   }
 };
 
-export let commentOnPost = async (
-  req: Request,
-  res: Response,
-  next: NextFunction
-) => {
+export let commentOnPost = async (req,res,next) => {
   const postId = req.body["postId"];
   const text = req.body["text"];
   console.log(req.body);
   db.collection("users")
-    .doc(req.body["userId"])
+    .doc(req.user.id)
     .get()
     .then(userDocSnap => {
       if (!userDocSnap.exists) {
@@ -275,7 +289,7 @@ export let commentOnPost = async (
           }
           let docData = docSnap.data();
           let comment = {
-            userId: req.body["userId"],
+            userId: req!.user.id,
             userName: userData.name,
             commentId: suid(),
             text: text,
@@ -303,11 +317,7 @@ export let commentOnPost = async (
     });
 };
 
-export let deleteComment = (
-  req: Request,
-  res: Response,
-  next: NextFunction
-) => {
+export let deleteComment = (req,res,next) => {
   let postId = req.body["postId"];
   let commentId = req.body["commentId"];
   db.collection(postCollection)
@@ -315,6 +325,9 @@ export let deleteComment = (
     .get()
     .then(docSnap => {
       let docData = docSnap.data();
+      if(docData.userId !== req.user.id){
+        res.status(401).send({message:"You are not authorized to delete this comment!"})
+      }
       docData.comment = docData.comment.filter(
         comment => comment.commentId !== commentId
       );
@@ -337,16 +350,12 @@ export let deleteComment = (
     });
 };
 
-export let replyOnComment = (
-  req: Request,
-  res: Response,
-  next: NextFunction
-) => {
+export let replyOnComment = (req,res,next) => {
   const postId = req.body["postId"];
   const commentId = req.body["commentId"];
   const text = req.body["text"];
   db.collection("users")
-    .doc(req.body["userId"])
+    .doc(req.user.id)
     .get()
     .then(userDocSnap => {
       if (!userDocSnap.exists) {
@@ -368,7 +377,7 @@ export let replyOnComment = (
                 comment.replies = [];
               }
               comment.replies.push({
-                userId: req.body["userId"],
+                userId: req.user.id,
                 userName: userData.name,
                 commentId: suid(),
                 text: text,
@@ -399,3 +408,37 @@ export let replyOnComment = (
         });
     });
 };
+
+
+// export let test = async (req,res,next) => {
+//   console.log("begg")  
+//   let post = {
+//     user_id: "Latest",
+//     imageUrl: "Latest",
+//     comment: [],
+//     likers: [],
+//     likesCount: 0,
+//     description: "Latest",
+//     Timestamp: new Date()
+//   };
+//   const newDoc = await db.collection(postCollection).add(post);
+//   const newId = newDoc.id;
+//   console.log("before")  
+//   res.status(200)
+//     .json({ message: "Post created successfully!", id: newDoc.id });
+//   console.log("after")  
+//   const userDoc = await db.collection("users").doc("CG3ziaR6LTpOyRtaeLXR").get();
+//   console.log(userDoc.exists)
+//   console.log(userDoc.data())
+//   const userConnectionsId = userDoc.data().connections;
+//   await db.collection(homeFeedCollection).doc("CG3ziaR6LTpOyRtaeLXR").collection("feed").doc(newId).set({
+//     ...post
+//   })
+//   userConnectionsId.forEach(id=>{
+//     db.collection("feed").doc(id).collection("feed").doc(newId).set({
+//       ...post
+//     })
+//     console.log(id)
+//   })
+//   return;
+// }
